+++ api/Models/Coin.cs
@@ -10,15 +10,15 @@ namespace api.Models
     public class Coin
     {
         public int Id { get; set; }
-        public string CoinName { get; set;}=string.Empty;
-        public string Symbol { get; set; } =string.Empty;
-        public decimal Price { get; set;}
+        public string CoinName { get; set; } = string.Empty;
+        public string Symbol { get; set; } = string.Empty;
+        public decimal Price { get; set; }
 
-        public long MarketCap {get; set;}
+        public decimal MarketCap { get; set; }
 
 
-        public List<Comment> Comments {get;set;} = new List<Comment>();
+        public List<Comment> Comments { get; set; } = new List<Comment>();
 
-        public List<Portfolio> Portfolios {get; set;} = new List<Portfolio>();
+        public List<Portfolio> Portfolios { get; set; } = new List<Portfolio>();
     }
 }
